# Don't hold DB mutex for block cache entry stat scans