--- a/cache/cache_entry_stats.h
+++ b/cache/cache_entry_stats.h
@@ -137,6 +137,7 @@ class CacheEntryStatsCollector {
                                  Cache::Priority::HIGH);
         if (!s.ok()) {
           assert(h == nullptr);
+          delete new_ptr;
           return s;
         }
       }