@pdillinger 2021-07-16
I haven't modified HISTORY.md because I intend for this to be picked into 6.21 and 6.22. (HISTORY would be updated in those for the patch.)